[Refactor] World.h
